- moved class TestFreeze from test.py to test_custom_schemas.py - moved classes TestDatasetSchemas and TestErrorMessages from test.py to test_type_checking.py - renamed what was left of test.py to test_json_pretty_viz.py - activated Travis for test_type_checking.py and test_json_pretty_viz.py - cleaned up some imports - renamed test_custom_operators.py to mock_custom_operators.py
