+++ test/test_custom_schemas.py
@@ -22,6 +22,7 @@ from xgboost import XGBClassifier as bar
 from lightgbm import LGBMClassifier as baz
 from sklearn.linear_model.least_angle import Lars as foobar
 from test.mock_module import UnknownOp
+from lale.search.lale_grid_search_cv import get_grid_search_parameter_grids
 
 
 class TestCustomSchema(unittest.TestCase):
@@ -331,3 +332,54 @@ class TestWrapUnknownOps(unittest.TestCase):
         cloned = clone(make_sklearn_compat(wrapped)).to_lale()
         self.assertTrue(isinstance(cloned, TrainableIndividualOp))
         self.assertEqual(cloned.hyperparams(), {'n_neighbors': 3})
+
+
+class TestFreeze(unittest.TestCase):
+    def test_individual_op_freeze_trainable(self):
+        from lale.lib.sklearn import LogisticRegression
+        liquid = LogisticRegression(C=0.1, solver='liblinear')
+        self.assertIn('penalty', liquid.free_hyperparams())
+        self.assertFalse(liquid.is_frozen_trainable())
+        liquid_grid = get_grid_search_parameter_grids(liquid)
+        self.assertTrue(len(liquid_grid) > 1, f'grid size {len(liquid_grid)}')
+        frozen = liquid.freeze_trainable()
+        self.assertEqual(len(frozen.free_hyperparams()), 0)
+        self.assertTrue(frozen.is_frozen_trainable())
+        frozen_grid = get_grid_search_parameter_grids(frozen)
+        self.assertEqual(len(frozen_grid), 1)
+
+    def test_pipeline_freeze_trainable(self):
+        from lale.lib.sklearn import PCA, LogisticRegression
+        liquid = PCA() >> LogisticRegression()
+        self.assertFalse(liquid.is_frozen_trainable())
+        liquid_grid = get_grid_search_parameter_grids(liquid)
+        self.assertTrue(len(liquid_grid) > 1, f'grid size {len(liquid_grid)}')
+        frozen = liquid.freeze_trainable()
+        self.assertTrue(frozen.is_frozen_trainable())
+        frozen_grid = get_grid_search_parameter_grids(frozen)
+        self.assertEqual(len(frozen_grid), 1)
+
+    def test_individual_op_freeze_trained(self):
+        from lale.lib.sklearn import KNeighborsClassifier
+        trainable = KNeighborsClassifier(n_neighbors=1)
+        X = [[0], [1], [2]]
+        y_old = [0, 0, 1]
+        y_new = [1, 0, 0]
+        liquid_old = trainable.fit(X, y_old)
+        self.assertEqual(list(liquid_old.predict(X)), y_old)
+        liquid_new = liquid_old.fit(X, y_new)
+        self.assertEqual(list(liquid_new.predict(X)), y_new)
+        frozen_old = trainable.fit(X, y_old).freeze_trained()
+        self.assertFalse(liquid_old.is_frozen_trained())
+        self.assertTrue(frozen_old.is_frozen_trained())
+        self.assertEqual(list(frozen_old.predict(X)), y_old)
+        frozen_new = frozen_old.fit(X, y_new)
+        self.assertEqual(list(frozen_new.predict(X)), y_old)
+
+    def test_pipeline_freeze_trained(self):
+        from lale.lib.sklearn import MinMaxScaler, LogisticRegression
+        trainable = MinMaxScaler() >> LogisticRegression()
+        liquid = trainable.fit([[0], [1], [2]], [0, 0, 1])
+        frozen = liquid.freeze_trained()
+        self.assertFalse(liquid.is_frozen_trained())
+        self.assertTrue(frozen.is_frozen_trained())
