+++ test/test_type_checking.py
@@ -13,141 +13,16 @@
 # limitations under the License.
 
 import unittest
-import warnings
-import random
 import jsonschema
-import sys
 
 import lale.lib.lale
 from lale.lib.lale import ConcatFeatures
 from lale.lib.lale import NoOp
-from lale.lib.sklearn import KNeighborsClassifier
-from lale.lib.sklearn import LinearSVC
 from lale.lib.sklearn import LogisticRegression
-from lale.lib.sklearn import MinMaxScaler
-from lale.lib.sklearn import MLPClassifier
-from lale.lib.sklearn import Nystroem
-from lale.lib.sklearn import OneHotEncoder
-from lale.lib.sklearn import PCA
 from lale.lib.sklearn import TfidfVectorizer
-from lale.lib.sklearn import MultinomialNB
-from lale.lib.sklearn import SimpleImputer
-from lale.lib.sklearn import SVC
-from lale.lib.xgboost import XGBClassifier
-from lale.lib.sklearn import PassiveAggressiveClassifier
-from lale.lib.sklearn import StandardScaler
-from lale.lib.sklearn import FeatureAgglomeration
 from lale.lib.sklearn import NMF
-from typing import List
-from lale.helpers import SubschemaError
 
 import sklearn.datasets
-
-from lale.sklearn_compat import make_sklearn_compat
-from lale.search.lale_grid_search_cv import get_grid_search_parameter_grids
-from lale.search.lale_smac import get_smac_space, lale_trainable_op_from_config
-
-class TestToJson(unittest.TestCase):
-    def test_with_operator_choice(self):
-        from lale.operators import make_union, make_choice, make_pipeline
-        kernel_tfm_or_not =  NoOp | Nystroem
-        tfm = PCA
-        clf = make_choice(LogisticRegression, KNeighborsClassifier)
-        clf.to_json()
-        optimizable = kernel_tfm_or_not >> tfm >> clf
-        optimizable.to_json()
-
-class TestToGraphviz(unittest.TestCase):
-    def test_with_operator_choice(self):
-        from lale.operators import make_union, make_choice, make_pipeline
-        from lale.helpers import to_graphviz
-        kernel_tfm_or_not =  NoOp | Nystroem
-        tfm = PCA
-        clf = make_choice(LogisticRegression, KNeighborsClassifier)
-        to_graphviz(clf)
-        optimizable = kernel_tfm_or_not >> tfm >> clf
-        to_graphviz(optimizable)
-
-    def test_invalid_input(self):
-        from sklearn.linear_model import LogisticRegression as SklearnLR
-        scikit_lr = SklearnLR()
-        from lale.helpers import to_graphviz
-        with self.assertRaises(ValueError):
-            to_graphviz(scikit_lr)
-
-class TestPrettyPrint(unittest.TestCase):
-    def round_trip(self, string1):
-        globals1 = {}
-        exec(string1, globals1)
-        pipeline1 = globals1['pipeline']
-        from lale.pretty_print import to_string
-        string2 = to_string(pipeline1)
-        self.maxDiff = None
-        self.assertEqual(string1, string2)
-        globals2 = {}
-        exec(string2, globals2)
-        pipeline2 = globals2['pipeline']
-
-    def test_reducible(self):
-        string1 = \
-"""from lale.lib.sklearn import MinMaxScaler
-from lale.lib.lale import NoOp
-from lale.lib.sklearn import PCA
-from lale.lib.sklearn import Nystroem
-from lale.lib.lale import ConcatFeatures
-from lale.lib.sklearn import KNeighborsClassifier
-from lale.lib.sklearn import LogisticRegression
-pca = PCA(copy=False)
-logistic_regression = LogisticRegression(solver='saga', C=0.9)
-pipeline = (MinMaxScaler | NoOp) >> (pca & Nystroem) >> ConcatFeatures >> (KNeighborsClassifier | logistic_regression)"""
-        self.round_trip(string1)
-
-    def test_import_as(self):
-        #code to reproduce in printing
-        from lale.lib.sklearn import MinMaxScaler as Scaler
-        from lale.lib.lale import NoOp
-        from lale.lib.sklearn import PCA
-        from lale.lib.sklearn import Nystroem
-        from lale.lib.lale import ConcatFeatures as Concat
-        from lale.lib.sklearn import KNeighborsClassifier as KNN
-        from lale.lib.sklearn import LogisticRegression as LR
-        pca = PCA(copy=False)
-        lr = LR(solver='saga', C=0.9)
-        pipeline = (Scaler | NoOp) >> (pca & Nystroem) >> Concat >> (KNN | lr)
-        #expected string
-        string1 = \
-"""from lale.lib.sklearn import MinMaxScaler as Scaler
-from lale.lib.lale import NoOp
-from lale.lib.sklearn import PCA
-from lale.lib.sklearn import Nystroem
-from lale.lib.lale import ConcatFeatures as Concat
-from lale.lib.sklearn import KNeighborsClassifier as KNN
-from lale.lib.sklearn import LogisticRegression as LR
-pca = PCA(copy=False)
-lr = LR(solver='saga', C=0.9)
-pipeline = (Scaler | NoOp) >> (pca & Nystroem) >> Concat >> (KNN | lr)"""
-        #testing harness
-        from lale.pretty_print import to_string
-        string2 = to_string(pipeline)
-        self.maxDiff = None
-        self.assertEqual(string1, string2)
-        globals2 = {}
-        exec(string2, globals2)
-        pipeline2 = globals2['pipeline']
-
-    def test_irreducible(self):
-        string1 = \
-"""from lale.lib.sklearn import PCA
-from lale.lib.sklearn import Nystroem
-from lale.lib.sklearn import MinMaxScaler
-from lale.lib.sklearn import LogisticRegression
-from lale.lib.sklearn import KNeighborsClassifier
-from lale.operators import get_pipeline_of_applicable_type
-step_1 = PCA | Nystroem
-pipeline = get_pipeline_of_applicable_type(
-    steps=[step_1, MinMaxScaler, LogisticRegression, KNeighborsClassifier],
-    edges=[(step_1,LogisticRegression), (MinMaxScaler,LogisticRegression), (MinMaxScaler,KNeighborsClassifier)])"""
-        self.round_trip(string1)
 
 class TestDatasetSchemas(unittest.TestCase):
     @classmethod
@@ -500,182 +375,3 @@ class TestErrorMessages(unittest.TestCase):
             LogisticRegression(solver='sag', penalty='l1')
         summary = cm.exception.message.split('\n')[0]
         self.assertEqual(summary, "Invalid configuration for LogisticRegression(solver='sag', penalty='l1') due to constraint the newton-cg, sag, and lbfgs solvers support only l2 penalties.")
-
-class TestFreeze(unittest.TestCase):
-    def test_individual_op_freeze_trainable(self):
-        liquid = LogisticRegression(C=0.1, solver='liblinear')
-        self.assertIn('penalty', liquid.free_hyperparams())
-        self.assertFalse(liquid.is_frozen_trainable())
-        liquid_grid = get_grid_search_parameter_grids(liquid)
-        self.assertTrue(len(liquid_grid) > 1, f'grid size {len(liquid_grid)}')
-        frozen = liquid.freeze_trainable()
-        self.assertEqual(len(frozen.free_hyperparams()), 0)
-        self.assertTrue(frozen.is_frozen_trainable())
-        frozen_grid = get_grid_search_parameter_grids(frozen)
-        self.assertEqual(len(frozen_grid), 1)
-
-    def test_pipeline_freeze_trainable(self):
-        liquid = PCA() >> LogisticRegression()
-        self.assertFalse(liquid.is_frozen_trainable())
-        liquid_grid = get_grid_search_parameter_grids(liquid)
-        self.assertTrue(len(liquid_grid) > 1, f'grid size {len(liquid_grid)}')
-        frozen = liquid.freeze_trainable()
-        self.assertTrue(frozen.is_frozen_trainable())
-        frozen_grid = get_grid_search_parameter_grids(frozen)
-        self.assertEqual(len(frozen_grid), 1)
-
-    def test_individual_op_freeze_trained(self):
-        trainable = KNeighborsClassifier(n_neighbors=1)
-        X = [[0], [1], [2]]
-        y_old = [0, 0, 1]
-        y_new = [1, 0, 0]
-        liquid_old = trainable.fit(X, y_old)
-        self.assertEqual(list(liquid_old.predict(X)), y_old)
-        liquid_new = liquid_old.fit(X, y_new)
-        self.assertEqual(list(liquid_new.predict(X)), y_new)
-        frozen_old = trainable.fit(X, y_old).freeze_trained()
-        self.assertFalse(liquid_old.is_frozen_trained())
-        self.assertTrue(frozen_old.is_frozen_trained())
-        self.assertEqual(list(frozen_old.predict(X)), y_old)
-        frozen_new = frozen_old.fit(X, y_new)
-        self.assertEqual(list(frozen_new.predict(X)), y_old)
-
-    def test_pipeline_freeze_trained(self):
-        trainable = MinMaxScaler() >> LogisticRegression()
-        liquid = trainable.fit([[0], [1], [2]], [0, 0, 1])
-        frozen = liquid.freeze_trained()
-        self.assertFalse(liquid.is_frozen_trained())
-        self.assertTrue(frozen.is_frozen_trained())
-
-class TestToAndFromJSON(unittest.TestCase):
-    def test_trainable_individual_op(self):
-        self.maxDiff = None
-        from lale.json_operator import to_json, from_json
-        from lale.lib.sklearn import LogisticRegression as LR
-        operator = LR(LR.solver.sag, C=0.1)
-        json_expected = {
-            'class': 'lale.lib.sklearn.logistic_regression.LogisticRegressionImpl',
-            'state': 'trainable',
-            'operator': 'LogisticRegression', 'label': 'LR', 'id': 'lr',
-            'documentation_url': 'http://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html',
-            'hyperparams': {'C': 0.1, 'solver': 'sag'},
-            'is_frozen_trainable': False}
-        json = to_json(operator)
-        self.assertEqual(json, json_expected)
-        operator_2 = from_json(json)
-        json_2 = to_json(operator_2)
-        self.assertEqual(json_2, json_expected)
-
-    def test_operator_choice(self):
-        self.maxDiff = None
-        from lale.json_operator import to_json, from_json
-        from lale.lib.sklearn import MinMaxScaler as Scl
-        operator = PCA | Scl
-        json_expected = {
-            'class': 'lale.operators.OperatorChoice',
-            'operator': 'OperatorChoice', 'id': 'choice',
-            'state': 'planned',
-            'steps': [
-            {   'class': 'lale.lib.sklearn.pca.PCAImpl',
-                'state': 'planned',
-                'operator': 'PCA', 'label': 'PCA', 'id': 'pca',
-                'documentation_url': 'https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html'},
-            {   'class': 'lale.lib.sklearn.min_max_scaler.MinMaxScalerImpl',
-                'state': 'planned',
-                'operator': 'MinMaxScaler', 'label': 'Scl', 'id': 'scl',
-                'documentation_url': 'https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.MinMaxScaler.html'}]}
-        json = to_json(operator)
-        self.assertEqual(json, json_expected)
-        operator_2 = from_json(json)
-        json_2 = to_json(operator_2)
-        self.assertEqual(json_2, json_expected)
-
-    def test_pipeline(self):
-        self.maxDiff = None
-        from lale.json_operator import to_json, from_json
-        from lale.lib.sklearn import LogisticRegression as LR
-        operator = (PCA & NoOp) >> ConcatFeatures >> LR
-        json_expected = {
-          'class': 'lale.operators.PlannedPipeline',
-          'state': 'planned',
-          'id': 'pipeline',
-          'edges': [[0, 2], [1, 2], [2, 3]],
-          'steps': [
-          { 'class': 'lale.lib.sklearn.pca.PCAImpl',
-            'state': 'planned',
-            'operator': 'PCA', 'label': 'PCA', 'id': 'pca',
-            'documentation_url': 'https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html'},
-          { 'class': 'lale.lib.lale.no_op.NoOpImpl',
-            'state': 'trained',
-            'operator': 'NoOp', 'label': 'NoOp', 'id': 'no_op',
-            'documentation_url': 'https://lale.readthedocs.io/en/latest/modules/lale.lib.lale.no_op.html',
-            'hyperparams': None,
-            'coefs': None,
-            'is_frozen_trainable': True, 'is_frozen_trained': True},
-          { 'class': 'lale.lib.lale.concat_features.ConcatFeaturesImpl',
-            'state': 'trained',
-            'operator': 'ConcatFeatures', 'label': 'ConcatFeatures', 'id': 'concat_features',
-            'documentation_url': 'https://lale.readthedocs.io/en/latest/modules/lale.lib.lale.concat_features.html',
-            'hyperparams': None,
-            'coefs': None,
-            'is_frozen_trainable': True, 'is_frozen_trained': True},
-          { 'class': 'lale.lib.sklearn.logistic_regression.LogisticRegressionImpl',
-            'state': 'planned',
-            'operator': 'LogisticRegression', 'label': 'LR', 'id': 'lr',
-            'documentation_url': 'http://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html'}]}
-        json = to_json(operator)
-        self.assertEqual(json, json_expected)
-        operator_2 = from_json(json)
-        json_2 = to_json(operator_2)
-        self.assertEqual(json, json_2)
-
-    def test_nested(self):
-        self.maxDiff = None
-        from lale.json_operator import to_json, from_json
-        from lale.lib.sklearn import LogisticRegression as LR
-        operator = PCA >> (LR(C=0.09) | NoOp >> LR(C=0.19))
-        json_expected = {
-          'class': 'lale.operators.PlannedPipeline',
-          'state': 'planned',
-          'id': 'pipeline_0',
-          'edges': [[0, 1]],
-          'steps': [
-          { 'class': 'lale.lib.sklearn.pca.PCAImpl',
-            'state': 'planned',
-            'operator': 'PCA', 'label': 'PCA', 'id': 'pca',
-            'documentation_url': 'https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html'},
-          { 'class': 'lale.operators.OperatorChoice',
-            'state': 'planned',
-            'operator': 'OperatorChoice', 'id': 'choice',
-            'steps': [
-            { 'class': 'lale.lib.sklearn.logistic_regression.LogisticRegressionImpl',
-              'state': 'trainable',
-              'operator': 'LogisticRegression', 'label': 'LR', 'id': 'lr_0',
-              'documentation_url': 'http://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html',
-              'hyperparams': {'C': 0.09},
-              'is_frozen_trainable': False},
-            { 'class': 'lale.operators.TrainablePipeline',
-              'state': 'trainable', 'id': 'pipeline_1',
-              'edges': [[0, 1]],
-              'steps': [
-              { 'class': 'lale.lib.lale.no_op.NoOpImpl',
-                'state': 'trained',
-                'operator': 'NoOp', 'label': 'NoOp', 'id': 'no_op',
-                'documentation_url': 'https://lale.readthedocs.io/en/latest/modules/lale.lib.lale.no_op.html',
-                'hyperparams': None,
-                'coefs': None,
-                'is_frozen_trainable': True, 'is_frozen_trained': True},
-              { 'class': 'lale.lib.sklearn.logistic_regression.LogisticRegressionImpl',
-                'state': 'trainable',
-                'operator': 'LogisticRegression', 'label': 'LR', 'id': 'lr_1',
-                'documentation_url': 'http://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html',
-                'hyperparams': {'C': 0.19},
-                'is_frozen_trainable': False}]}]}]}
-        json = to_json(operator)
-        self.assertEqual(json, json_expected)
-        operator_2 = from_json(json)
-        json_2 = to_json(operator_2)
-        self.assertEqual(json, json_2)
-
-if __name__ == '__main__':
-    unittest.main()
